Move the vp wrappers into scripts/vp, and trim their AGENTS.md notes
`git mv` only, plus the references that follow:

- scripts/vp-run.ts        -> scripts/vp/run.ts
- scripts/vp-concurrency.ts -> scripts/vp/concurrency.ts (+ its test)

The prefix drops because the directory now carries it, matching
`preview/staging-config.ts` rather than `release/build-release.ts` -- and
`node scripts/vp/run.ts` reads better than the stutter.

Two path computations needed a third `dirname` hop, since both walked up to the
repo root from `scripts/`: `run.ts`'s `ROOT` (which locates
`node_modules/vite-plus/bin/vp`, a *root* devDependency) and `concurrency.ts`'s
`ROOT_ENV_FILE`. The first would have silently pointed into `scripts/node_modules`,
which exists as a workspace package but has no vite-plus.

`ROOT_ENV_FILE`'s test asserted `join(dirname(import.meta.dirname), ".env")`,
mirroring the same hop count from the same directory -- so it would have gone
green alongside the bug it exists to catch, which is exactly what happened on the
first pass of this move. It now pins the resolved directory against
`pnpm-workspace.yaml`, a file that only exists at the workspace root, and fails
with "scripts is not the workspace root" if the hops are ever wrong again.

Everything else was recursive globs (`scripts/**/*.test.ts`, the scripts
tsconfig `include`, the oxlint overrides) or `readdirSync` walks
(env-passthrough), so nothing there needed touching.

While here, trimmed what this PR adds to AGENTS.md by ~47% (3566 -> 1892 chars),
cutting the history rather than the rules: how `build`/`clean` used to get away
with `-r`, the `--no-cache` experiment that proved pnpm appends forwarded args,
the per-source note-printing rationale, and a `-r` warning the new filter bullet
now states more fully. The traps stay -- spell the filter `=`-joined and
unquoted for cmd.exe, `pnpm build --concurrency-limit 2` does not reach vp, use
the env var or the root `.env` instead.

Verified: 266/266 scripts tests, `pnpm build`/`test`/`lint` all exit 0 with no
`scripts/` diagnostics, the concurrency note still prints through the new path,
and a root `.env` holding a limit is still found (`concurrency 3 (from .env)`).
