+++ package.json
@@ -5,9 +5,9 @@
   "type": "module",
   "packageManager": "pnpm@11.17.0+sha512.cca3cea332ad254bb84145f966d19f4879615210346fc92c79a047f23a0d7b3cca3c3792f0076ba1f1831d277efbcf0a9119b31a9a60eca7fb3d6231f331ef72",
   "scripts": {
-    "build": "node scripts/vp-run.ts --filter=!cloudflare-os --cache build",
+    "build": "node scripts/vp/run.ts --filter=!cloudflare-os --cache build",
     "run-local": "node scripts/run-local.ts",
-    "test": "node scripts/vp-run.ts --filter=!cloudflare-os --cache test",
+    "test": "node scripts/vp/run.ts --filter=!cloudflare-os --cache test",
     "evals": "pnpm --filter @gadgets/workshop-evals evals",
     "evals:ui": "pnpm --filter @gadgets/workshop-evals evals:ui",
     "preview:config": "node scripts/preview/preview.ts config",
@@ -16,7 +16,7 @@
     "preview:sweep": "node scripts/preview/preview.ts sweep",
     "dev-client": "cd packages/workshop-frontend && pnpm run dev",
     "dev-server": "node scripts/run-dev-server.ts",
-    "clean": "node scripts/vp-run.ts --filter=!cloudflare-os clean",
+    "clean": "node scripts/vp/run.ts --filter=!cloudflare-os clean",
     "lint:check": "vp lint",
     "lint:fix": "vp lint --fix",
     "types:check": "pnpm run build",
