+++ scripts/env-passthrough.test.ts
@@ -91,7 +91,7 @@ const EXPECTED: Record<string, ExpectedArea> = {
       "GITHUB_REPOSITORY", "GITHUB_TOKEN", "PREVIEW_ADMINS", "PREVIEW_GITHUB_CLIENT_ID",
       "PREVIEW_GITHUB_CLIENT_SECRET", "PREVIEW_NAME", "PREVIEW_PR_NUMBER",
       "PREVIEW_WORKERS_DEV_HOST", "PREVIEW_WRANGLER", "VITE_BACKEND_HOST",
-      // Read by `vp-concurrency.ts` in the wrapper before `vp` starts, never inside a task.
+      // Read by `vp/concurrency.ts` in the wrapper before `vp` starts, never inside a task.
       "VP_RUN_CONCURRENCY_LIMIT",
     ],
   },
